calibration functions are done

+++ i2c_mpu6050.hpp
@@ -7,12 +7,12 @@ class i2c_mpu6050{
 private:
 	hwlib::i2c_bus_bit_banged_scl_sda & bus;
 	uint8_t chipAddr;
-	uint16_t acXoffset = 0x00;
-	uint16_t acYoffset = 0x00;
-	uint32_t acZoffset = 0x00;
-	uint16_t GyXoffset = 0x00;
-	uint16_t GyYoffset = 0x00;
-	uint16_t GyZoffset = 0x00;
+	int16_t acXoffset = 0x00;
+	int16_t acYoffset = 0x00;
+	int16_t acZoffset = 0x00;
+	int16_t GyXoffset = 0x00;
+	int16_t GyYoffset = 0x00;
+	int16_t GyZoffset = 0x00;
 	
 	void initialize();
 	void header_values();
@@ -28,14 +28,14 @@ public:
 	
 	void set_register(const uint8_t &regAddr, const uint8_t & data);
 	
-	uint16_t read_values(const uint8_t &regAddr);
-	uint8_t read_word(const uint8_t &regAddr);
+	int16_t read_values(const uint8_t &regAddr);
+	int16_t read_word(const uint8_t &regAddr);
 	
 	void display_raw_values();
 	void display_values();
 	void display_rpy(); //rpy is short for roll - pitch and yaw.
-	void set_calibrate_values(const uint16_t & acXoffset, const uint16_t & acYoffset, const uint16_t & acZoffset, const uint16_t & GyXoffset,
-	const uint16_t & GyYoffset, const uint16_t & GyZoffset);
+	void set_calibrate_values(const int16_t & acXoffset, const int16_t & acYoffset, const int16_t & acZoffset, const int16_t & GyXoffset,
+	const int16_t & GyYoffset, const int16_t & GyZoffset);
 	void calibrate();
 	void show_calibrate_values();
 };
